requirements.txt
openpyxl>=3.1.0
pyyaml>=6.0
# pymodbus is NOT required — Modbus TCP server is implemented in pure Python
# Install pymodbus if you want to use it as an alternative client for testing:
#   pip install pymodbus
